no need 64bits, specially for casts

+++ src/bn/relic_bn_mxp.c
@@ -386,7 +386,7 @@ void _bn_mxp_sim(bn_t S, const bn_t P[BN_XPWDT], bn_t u[BN_XPWDT],
                  const bn_t T[BN_XPWDT], const bn_t mod) {
         // WARNING: overwrites u
     int iszeroexp = 0x1;
-    for(uint64_t j=0; j<BN_XPWDT; ++j) {
+    for(unsigned int j=0; j<BN_XPWDT; ++j) {
         iszeroexp &= bn_is_zero(u[j]);
     }
 
@@ -396,11 +396,11 @@ void _bn_mxp_sim(bn_t S, const bn_t P[BN_XPWDT], bn_t u[BN_XPWDT],
     }
 
         // Select odd exponents
-    uint64_t parities = !bn_is_even(u[0]);
-    for(uint64_t j=1; j<BN_XPWDT; ++j) parities |= (!bn_is_even(u[j]))<<j;
+    unsigned int parities = !bn_is_even(u[0]);
+    for(unsigned int j=1; j<BN_XPWDT; ++j) parities |= (unsigned int)(!bn_is_even(u[j]))<<j;
 
         // Halving exponents
-    for(uint64_t j=0; j<BN_XPWDT; ++j) {
+    for(unsigned int j=0; j<BN_XPWDT; ++j) {
         bn_hlv(u[j],u[j]);			// WARNING: u overwriten
     }
 
@@ -421,7 +421,7 @@ void _bn_mxp_sim(bn_t S, const bn_t P[BN_XPWDT], bn_t u[BN_XPWDT],
  * bn_t simultaneous exponentiation generalized Shamir trick and fixed width
  ****************************************************************/
 void bn_mxp_sim(bn_t S, const bn_t P[BN_XPWDT], const bn_t u[BN_XPWDT], const bn_t mod) {
-    const unsigned int TABLE_DIM = 1<<BN_XPWDT;
+    const unsigned int TABLE_DIM = 1u<<BN_XPWDT;
     bn_t T[TABLE_DIM], hu[BN_XPWDT];
 
     RLC_TRY {
@@ -429,17 +429,17 @@ void bn_mxp_sim(bn_t S, const bn_t P[BN_XPWDT], const bn_t u[BN_XPWDT], const bn
         bn_null(T[0]); bn_new(T[0]); bn_set_dig(T[0],1);
 
             // Precompute all 2^{BN_XPWDT} combinations of points P
-        for(uint64_t i=0; i<BN_XPWDT; ++i) {
-            const uint64_t star = 1<<i; const uint64_t stars = star<<1;
+        for(unsigned int i=0; i<BN_XPWDT; ++i) {
+            const unsigned int star = 1<<i; const unsigned int stars = star<<1;
             bn_null(T[star]); bn_new(T[star]); bn_copy(T[star], P[i]);
-            for(uint64_t j=star+1; j<stars; ++j) {
+            for(unsigned int j=star+1; j<stars; ++j) {
                 bn_null(T[j]); bn_new(T[j]);
                 bn_mul(T[j], T[star], T[j-star]); bn_mod(T[j],T[j],mod);
             }
         }
 
             // copy u, as hu will be overwritten by the subroutine
-        for(uint64_t j=0; j<BN_XPWDT; ++j) {
+        for(unsigned int j=0; j<BN_XPWDT; ++j) {
             bn_null(hu[j]); bn_new(hu[j]); bn_copy(hu[j],u[j]);
         }
 
@@ -449,8 +449,8 @@ void bn_mxp_sim(bn_t S, const bn_t P[BN_XPWDT], const bn_t u[BN_XPWDT], const bn
 	} RLC_CATCH_ANY {
 		RLC_THROW(ERR_CAUGHT);
 	} RLC_FINALLY {
-        for(uint64_t i=0; i<TABLE_DIM; ++i) { bn_free(T[i]); }
-        for(uint64_t j=0; j<BN_XPWDT; ++j) { bn_free(hu[j]); }
+        for(unsigned int i=0; i<TABLE_DIM; ++i) { bn_free(T[i]); }
+        for(unsigned int j=0; j<BN_XPWDT; ++j) { bn_free(hu[j]); }
     }
 }
 
@@ -462,7 +462,7 @@ void bn_mxp_sim_lot(bn_t S, const bn_t P[], const bn_t u[], const bn_t mod, int 
 	RLC_TRY {
             // Will use blocks of size BN_XPWDT
         bn_null(tmp); bn_new(tmp);
-        for(uint64_t j=0; j<BN_XPWDT; ++j) {
+        for(unsigned int j=0; j<BN_XPWDT; ++j) {
             bn_null(wP[j]); bn_new(wP[j]);
             bn_null(wu[j]); bn_new(wu[j]);
         }
@@ -472,7 +472,7 @@ void bn_mxp_sim_lot(bn_t S, const bn_t P[], const bn_t u[], const bn_t mod, int 
         bn_set_dig(S, 1);
             // Exponentiate by blocks of size BN_XPWDT
         int i = 0; for(; i<endblockingloop; ) {
-            for(uint64_t j=0; j<BN_XPWDT; ++j, ++i) {
+            for(unsigned int j=0; j<BN_XPWDT; ++j, ++i) {
                 bn_copy(wP[j], P[i]);
                 bn_copy(wu[j], u[i]);
             }
@@ -491,7 +491,7 @@ void bn_mxp_sim_lot(bn_t S, const bn_t P[], const bn_t u[], const bn_t mod, int 
 	} RLC_CATCH_ANY {
 		RLC_THROW(ERR_CAUGHT);
 	} RLC_FINALLY {
-        for(uint64_t j=0; j<BN_XPWDT; ++j) {
+        for(unsigned int j=0; j<BN_XPWDT; ++j) {
             bn_free(wP[j]); bn_free(wu[j]);
         }
         bn_free(tmp);
